Improve logcontext handling for @cache
Rather than clearing the logcontext when we get a cache miss, pass it through
to the DB transaction method, but clear it before storing the deferreds in the
cache.

This will mean that DB queries will get logged against the request which caused
the cache miss - which although isn't perfect seems better than treating them
as completely random transactions.